move quadrupole.py to plexscoring
