+++ alphadia/plotting/debug.py
@@ -6,7 +6,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from alphadia import quadrupole
+from alphadia.plexscoring import quadrupole
 from alphadia.plotting import utils
